Pushing quick text files used for testing
